refactor(admin): split writing actions
Move writing status, creation, and content update persistence out of the root admin server action registry so actions.ts stays focused on exported action boundaries.

Verified with pnpm --filter @anipotts/admin typecheck, pnpm --filter @anipotts/admin lint, pnpm --filter @anipotts/admin build, pnpm exec prettier --check apps/admin/src/app/actions.ts apps/admin/src/app/lib/content-records.ts, git diff --check -- apps/admin/src/app/actions.ts apps/admin/src/app/lib/content-records.ts, and pnpm dlx knip --no-exit-code.
